Moved common CommandSender based methods to OpSender out of NMS implementations
