+++ ConfigCommands-core/src/main/java/me/willkroboth/ConfigCommands/NMS/OpSender.java
@@ -2,6 +2,10 @@ package me.willkroboth.ConfigCommands.NMS;
 
 import me.willkroboth.ConfigCommands.ConfigCommandsHandler;
 import org.bukkit.command.CommandSender;
+import org.bukkit.permissions.Permission;
+
+import java.util.Arrays;
+import java.util.UUID;
 
 // OpSenders should:
 //  Appear as the class they represent in instanceof calls
@@ -22,4 +26,46 @@ public interface OpSender extends CommandSender {
 
     // provide result message
     String getResult();
+
+    // store result message for CommandSender methods
+    void setLastMessage(String message);
+
+    // overriding CommandSender
+    default void sendMessage(String s) {
+        setLastMessage(s);
+    }
+
+    default void sendMessage(String[] strings) {
+        setLastMessage(Arrays.toString(strings));
+    }
+
+    default void sendMessage(UUID uuid, String s) {
+        setLastMessage(s);
+    }
+
+    default void sendMessage(UUID uuid, String[] strings) {
+        setLastMessage(Arrays.toString(strings));
+    }
+
+    // overriding ServerOperator
+    default boolean isOp() {
+        return true;
+    }
+
+    // permissions
+    default boolean isPermissionSet(String name) {
+        return true;
+    }
+
+    default boolean isPermissionSet(Permission perm) {
+        return true;
+    }
+
+    default boolean hasPermission(String name) {
+        return true;
+    }
+
+    default boolean hasPermission(Permission perm) {
+        return true;
+    }
 }
