+++ ConfigCommands-nms/ConfigCommands-1_19/src/main/java/me/willkroboth/ConfigCommands/NMS/V1_19/OpSenders1_19/OpSender1_19.java
@@ -11,10 +11,6 @@ import org.bukkit.craftbukkit.v1_19_R1.command.CraftBlockCommandSender;
 import org.bukkit.craftbukkit.v1_19_R1.command.ProxiedNativeCommandSender;
 import org.bukkit.craftbukkit.v1_19_R1.entity.CraftMinecartCommand;
 import org.bukkit.craftbukkit.v1_19_R1.entity.CraftPlayer;
-import org.bukkit.permissions.Permission;
-
-import java.util.Arrays;
-import java.util.UUID;
 
 public interface OpSender1_19 extends OpSender, CommandSource {
     static OpSender1_19 makeOpSender(CommandSender sender) {
@@ -38,53 +34,6 @@ public interface OpSender1_19 extends OpSender, CommandSource {
 
     static CommandSourceStack modifyStack(CommandSourceStack source, OpSender1_19 sender) {
         return source.withSource(sender).withPermission(4);
-    }
-
-    CommandSender getSender();
-
-    // provide result message
-    String getResult();
-
-    // store result message for CommandSender methods
-    void setLastMessage(String message);
-
-    // overriding CommandSender
-    default void sendMessage(String s) {
-        setLastMessage(s);
-    }
-
-    default void sendMessage(String[] strings) {
-        setLastMessage(Arrays.toString(strings));
-    }
-
-    default void sendMessage(UUID uuid, String s) {
-        setLastMessage(s);
-    }
-
-    default void sendMessage(UUID uuid, String[] strings) {
-        setLastMessage(Arrays.toString(strings));
-    }
-
-    // overriding ServerOperator
-    default boolean isOp() {
-        return true;
-    }
-
-    // permissions
-    default boolean isPermissionSet(String name) {
-        return true;
-    }
-
-    default boolean isPermissionSet(Permission perm) {
-        return true;
-    }
-
-    default boolean hasPermission(String name) {
-        return true;
-    }
-
-    default boolean hasPermission(Permission perm) {
-        return true;
     }
 
     // CommandSource methods
